CPR-489-cpr-uuid refactor addUUID to seperate class
